move search result components
