--- ui/shared/components/panel/search/VariantSearchResults.test.js
@@ -1,16 +1,0 @@
-import React from 'react'
-import { shallow, configure } from 'enzyme'
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17'
-import configureStore from 'redux-mock-store'
-
-import VariantSearchResults from './VariantSearchResults'
-
-import { STATE } from '../fixtures'
-
-configure({ adapter: new Adapter() })
-
-test('shallow-render without crashing', () => {
-  const store = configureStore()(STATE)
-
-  shallow(<VariantSearchResults store={store} />)
-})
